app/requirements.txt
# Python dependencies for the Medical Appointments Management System
# ============================================================================

# Web Framework
Django>=4.2,<5.0
django-environ==0.11.2

# PostgreSQL database adapter
psycopg2-binary==2.9.9

# Database connection pooling (optional, for future optimization)
psycopg2-pool==1.1

# Environment variables management
python-dotenv==1.0.0

# Password hashing
bcrypt==4.1.2

# Logging and utilities
colorama==0.4.6

# Date and time utilities
python-dateutil==2.8.2

# Input validation
email-validator==2.1.0.post1

# Pretty printing for debugging
tabulate==0.9.0

# Rich terminal output
rich==13.7.0
